+++ src/main/java/insane96mcp/iguanatweaksexpanded/module/experience/enchantments/enchantment/GravityDefying.java
@@ -27,7 +27,7 @@ public class GravityDefying extends Enchantment implements IAttributeEnchantment
     }
 
     public static float getGravityReduction() {
-        return 0.5f;
+        return 0.45f;
     }
 
     public static void applyFallDamageReduction(LivingFallEvent event) {