+++ changelog.md
@@ -1,5 +1,7 @@
 # Changelog
 
+## Upcoming
+
 ## 1.8.2
 * Reach can now be applied to Chestplates
 * Gravity defying