+++ changelog.md
@@ -2,6 +2,9 @@
 
 ## Upcoming
 * Bees are now smaller
+* Gravity defying
+  * Reduced fall distance reduction (1.5 -> 0.5)
+  * Slightly reduced gravity reduction
 
 ## 1.8.2
 * Reach can now be applied to Chestplates